Separate non-components from components
